Update ViolationBuilder signature + Update Check interface breaking changes
== Check Interface ==
Instead of having one `run()` method, now we separated the getMatches with processMatch methods.

In a way, it's similar to PHPStan's Rule interface. getNodeTypes and processNodes.

The processMatch method returns an array of ViolationBuilder instances. If it returns an empty array, the match will not be counted as a violation.

The ViolationBuilder now changed. The identifier and file are not required. If they are not present, they will be added automatically by the CheckRunner. If they are present, they won't be changed.

== ViolationBuilder ==

The explanation is changed for `message` now. And the `build` method should not be called from the processMatch. It will be called from the CheckRunner, with the fallback identifier as the check name, and the fallback file as the match found initially.